[A4] Updated test-regression.sh to run multiple trials

+++ A4-raft-consensus/src/raft/test-regression.sh
@@ -15,37 +15,60 @@ tests=(
   TestRejoin
 )
 
-# Run tests
+#----------------------------
+
+# Check usage
+if [ $# -ne 1 ] || [ "$1" == "--help" ]; then
+  echo "Usage: ./test-regression.sh NUM-TRIALS"
+  exit
+fi
+
+trials=$1
+
+#----------------------------
+
+# Run trials
 pass="true"
-for test in ${tests[@]}
-do
-  echo "Running $test..."
-  outfile=out-$test.tmp
-  errfile=err-$test.tmp
+for trial in $(seq $trials); do
 
-  # Run test
-  go test -run $test > $outfile 2> $errfile
+  echo "TRIAL $trial"
 
-  # Check for failures and warnings
-  cat $outfile | grep "FAIL" > fail-out.tmp
-  cat $errfile | grep "FAIL" > fail-err.tmp
-  cat $outfile | grep "warning" > warning-out.tmp
-  cat $errfile | grep "warning" > warning-err.tmp
-  if [ -s fail-out.tmp ] || [ -s fail-err.tmp ]; then
-    pass="false"
-    echo "[!] Failed $test. Stdout in $outfile, stderr in $errfile"
-  elif [ -s warning-out.tmp ] || [ -s warning-err.tmp ]; then
-    echo "[!] Warning from $test. Stdout in $outfile, stderr in $errfile"
-  else
-    rm -f $outfile $errfile
+  # Run each test
+  for test in ${tests[@]}; do
+    echo "Running $test..."
+    outfile=out-$test-$trial.tmp
+    errfile=err-$test-$trial.tmp
+
+    # Run test
+    go test -run $test > $outfile 2> $errfile
+
+    # Check for failures and warnings
+    cat $outfile | grep "FAIL" > fail-out.tmp
+    cat $errfile | grep "FAIL" > fail-err.tmp
+    cat $outfile | grep "warning" > warning-out.tmp
+    cat $errfile | grep "warning" > warning-err.tmp
+    if [ -s fail-out.tmp ] || [ -s fail-err.tmp ]; then
+      pass="false"
+      echo "[!] Failed $test. Stdout in $outfile, stderr in $errfile"
+    elif [ -s warning-out.tmp ] || [ -s warning-err.tmp ]; then
+      echo "[!] Warning from $test. Stdout in $outfile, stderr in $errfile"
+    else
+      rm -f $outfile $errfile
+    fi
+
+    # Clean up
+    rm -rf fail-out.tmp fail-err.tmp warning-out.tmp warning-err.tmp
+  done
+
+  echo
+
+  if [ "$pass" == "false" ]; then
+    echo "FAIL during trial $trial"
+    break
   fi
 
-  # Clean up
-  rm -rf fail-out.tmp fail-err.tmp warning-out.tmp warning-err.tmp
 done
 
 if [ "$pass" == "true" ]; then
   echo "PASS"
-else
-  echo "FAIL"
 fi
